Extract getRequestBodyType
