+++ plugins/typescript/src/generators/generateReactQueryComponents.ts
@@ -8,20 +8,15 @@ import {
   OperationObject,
   ParameterObject,
   PathItemObject,
-  ReferenceObject,
-  RequestBodyObject,
 } from "openapi3-ts";
 import { get, groupBy } from "lodash";
 
-import {
-  getType,
-  schemaToTypeAliasDeclaration,
-} from "../core/schemaToTypeAliasDeclaration";
-import { findCompatibleMediaType } from "../core/findCompatibleMediaType";
+import { schemaToTypeAliasDeclaration } from "../core/schemaToTypeAliasDeclaration";
 import { getUsedImports } from "../core/getUsedImports";
 import { getVariablesType } from "../core/getVariablesType";
 import { paramsToSchema } from "../core/paramsToSchema";
 import { getResponseType } from "../core/getResponseType";
+import { getRequestBodyType } from "../core/getRequestBodyType";
 
 import { getCustomFetcher } from "../templates/customFetcher";
 import { getContext } from "../templates/context";
@@ -382,76 +377,6 @@ const getParamsGroupByType = (
   );
 
   return { queryParams, pathParams, headerParams };
-};
-
-/**
- * Extract types from request body
- */
-const getRequestBodyType = ({
-  requestBody,
-  components,
-}: {
-  requestBody?: RequestBodyObject | ReferenceObject;
-  components?: ComponentsObject;
-}) => {
-  if (!requestBody) {
-    return f.createKeywordTypeNode(ts.SyntaxKind.UndefinedKeyword);
-  }
-
-  if (isReferenceObject(requestBody)) {
-    const [hash, topLevel, namespace, name] = requestBody.$ref.split("/");
-    if (hash !== "#" || topLevel !== "components") {
-      throw new Error(
-        "This library only resolve $ref that are include into `#/components/*` for now"
-      );
-    }
-    if (namespace !== "requestBodies") {
-      throw new Error(
-        "$ref for requestBody must be on `#/components/requestBodies`"
-      );
-    }
-    return f.createTypeReferenceNode(
-      f.createQualifiedName(
-        f.createIdentifier("RequestBodies"),
-        f.createIdentifier(c.pascal(name))
-      ),
-      undefined
-    );
-  }
-
-  const mediaType = findCompatibleMediaType(requestBody);
-  if (!mediaType) {
-    return f.createKeywordTypeNode(ts.SyntaxKind.UndefinedKeyword);
-  }
-
-  if (isReferenceObject(mediaType)) {
-    const [hash, topLevel, namespace, name] = mediaType.$ref.split("/");
-    if (hash !== "#" || topLevel !== "components") {
-      throw new Error(
-        "This library only resolve $ref that are include into `#/components/*` for now"
-      );
-    }
-    if (namespace !== "schemas") {
-      throw new Error("$ref for schemas must be on `#/components/schemas`");
-    }
-
-    return f.createTypeReferenceNode(
-      f.createQualifiedName(
-        f.createIdentifier("Schemas"),
-        f.createIdentifier(c.pascal(name))
-      ),
-      undefined
-    );
-  }
-
-  if (!mediaType.schema) {
-    return f.createKeywordTypeNode(ts.SyntaxKind.UndefinedKeyword);
-  }
-
-  return getType(mediaType.schema, {
-    currentComponent: null,
-    openAPIDocument: { components },
-  });
 };
 
 /**
